+++ lib/src/hour.dart
@@ -6,7 +6,7 @@ import 'package:date/src/time_ordering.dart';
 import 'package:date/src/interval.dart';
 
 /// An immutable class to represent an hour.
-class Hour extends Interval implements TimeOrdering<Hour>, ComparableWithAdd<Hour> {
+class Hour extends Interval implements TimeOrdering<Hour>, Additive<Hour> {
   /// number of hours since origin in UTC timezone, hour beginning
   /// should explore if this is worth doing.
   ///int _value;
@@ -59,8 +59,6 @@ class Hour extends Interval implements TimeOrdering<Hour>, ComparableWithAdd<Hou
     Hour hour = other;
     return start == hour.start;
   }
-
-  int compareTo(Hour other) => this.start.compareTo(other.start);
 
   int get hashCode => start.hashCode;