+++ test/test_all.dart
@@ -10,12 +10,10 @@ import 'term_parse_test.dart' as termParseTest;
 import 'month_parse_test.dart' as monthParseTest;
 
 main() async {
-  Map env = Platform.environment;
-  String tzdb = env['HOME'] + '/.pub-cache/hosted/pub.dartlang.org/timezone-0.4.3/lib/data/2015b_all.tzf';
-  await initializeTimeZone(tzdb);
+  await initializeTimeZone();
 
   dateTest.testDate();
-  dateTest.testDateIterable();
+  //dateTest.testDateIterable();
 
   hourTest.hourTest();
   intervalTest.testInterval();